dydid

Set specifications for dynamic Difference-in-Differences analysis, report coefficients from regressions, and draw coefficients in a dynamic way.

Usage

dydid(
  data = df,
  dv = "y",
  tpoint = "reform_timepoint",
  step = 1,
  eid = "state_id",
  tid = "year_id",
  span = c(-5, 5),
  covar = NULL
)

Arguments

data

a data set object

dv

the outcome variable of your interest

tpoint

variable in your data that indicates the time when the treatment takes place

step

the unit by which the time ID increments, e.g., 2 for every two years

eid

ID for each individual entity

tid

ID for each time point

span

the number of periods before and after the treatment you want to take into consideration.

covar

controls variables to be included

Value

This functions return the following three objects: A. "regs": Original results of DiD regressions using felm(). B. "results": A data frame reporting coefficients and 95

C. "plot": A dynamic plot of DiD coefficients using ggplot2().

Examples

# load in your data
df <- read.dta13("data/big bad banks.dta")
#> Error in read.dta13("data/big bad banks.dta"): could not find function "read.dta13"

# use the function
bbb.did <- dydid(
  data = df, dv = "y", tpoint = "reform_timepoint", step = 1,
  eid = "state_id", tid = "year_id",
  span = c(-7, 7), covar = NULL
)
#> Error in df[, tid]: object of type 'closure' is not subsettable

# see the origigal regression results
bbb.did$regs
#> Error in eval(expr, envir, enclos): object 'bbb.did' not found

# see the organized table for coefficients and 95% CIs
bbb.did$results
#> Error in eval(expr, envir, enclos): object 'bbb.did' not found

# see the dynamic plot of DiD results
bbb.did$plot
#> Error in eval(expr, envir, enclos): object 'bbb.did' not found